fourth

+++ FlightUserPro/src/main/java/com/flightapp/Repository/BookTicketRepository.java
@@ -10,16 +10,16 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.flightapp.Entity.BookTicket;
 
-public interface BookTicketRepository  extends JpaRepository<BookTicket, Long> {
+public interface BookTicketRepository extends JpaRepository<BookTicket, Integer> {
 
 	@Query("SELECT BT FROM BookTicket BT WHERE BT.email LIKE ?1%")
-    List<BookTicket> searchByEmail(String email);
+	List<BookTicket> searchByEmail(String email);
 
-    @Query("SELECT BT FROM BookTicket BT WHERE BT.pnrNo LIKE ?1%")
-    List<BookTicket> searchByPnr(String pnrNo);
+	@Query("SELECT BT FROM BookTicket BT WHERE BT.pnrNo LIKE ?1%")
+	List<BookTicket> searchByPnr(String pnrNo);
 
-    @Transactional
-    @Modifying
-    @Query("UPDATE BookTicket BT SET BT.status = :status WHERE BT.pnrNo = :pnrNo")
-    int cancelTicket(String pnrNo, String status);
+	@Transactional	
+	@Modifying
+	@Query("UPDATE BookTicket BT SET BT.status = :status WHERE BT.pnrNo = :pnrNo")
+	int cancelTicket(String pnrNo, String status);
 }
